+++ idp_engine/Expression.py
@@ -90,6 +90,9 @@ class ASTNode(object):
     def interpret(self, problem: Any) -> "Expression":
         return self  # monkey-patched
 
+    def EN(self):
+        pass  # monkey-patched
+
 
 class Annotations(ASTNode):
     def __init__(self, **kwargs):
@@ -963,17 +966,17 @@ class AAggregate(Expression):
         self.q = ''
         super().__init__()
 
-
     def __str1__(self):
         if not self.annotated:
-            vars = "".join([f"{q}" for q in self.quantees])
-            out = ((f"{self.aggtype}(lambda {vars} : "
-                    f"{self.sub_exprs[0].str}"
-                    f")" ) if self.aggtype != "#" else
-                   (f"{self.aggtype}{{{vars}: "
-                    f"{self.sub_exprs[0].str}"
-                    f"}}")
-            )
+            vars = ",".join([f"{q}" for q in self.quantees])
+            if self.aggtype in ["sum", "min", "max"]:
+                out = (f"{self.aggtype}(lambda {vars} : "
+                        f"{self.sub_exprs[0].str}"
+                        f")" )
+            else:
+                out = (f"{self.aggtype}{{{vars} : "
+                       f"{self.sub_exprs[0].str}"
+                       f"}}")
         else:
             out = (f"{self.aggtype}{{"
                    f"{','.join(e.str for e in self.sub_exprs)}"
@@ -1000,7 +1003,7 @@ class AppliedSymbol(Expression):
     """Represents a symbol applied to arguments
 
     Args:
-        symbol (Expression): the symbol to be applied to arguments
+        symbol (SymbolExpr): the symbol to be applied to arguments
 
         is_enumerated (string): '' or 'is enumerated' or 'is not enumerated'