+++ idp_engine/Theory.py
@@ -320,9 +320,6 @@ class Theory(object):
             if s.code not in self.assignments:
                 self.assignments.assert__(s, None, S.UNKNOWN)
 
-        for ass in self.assignments.values():
-            ass.sentence.original = ass.sentence
-
         self._constraintz = None
         return self